refactor(mcp): rename oauth/ to outbound_credentials/
The subdomain resolves every egress credential mode (api_key, none, aws_sigv4, passthrough),
not just OAuth grants, so outbound_credentials names it accurately. Pure rename: relative
imports are unaffected; the spike import, the absolute test imports, and the path mentions in
the module/README docstrings are updated. The RFC 8693 token-type URN is left untouched.
