+++ litellm/proxy/gateway/mcp/__init__.py
@@ -4,7 +4,7 @@ Written v2-native: nothing here imports from v1 (the rest of
 `litellm.proxy._experimental.mcp_server.*`). v1 only ever reaches v2 through a thin
 adapter built in Phase 1, never the other way round.
 
-Phase 0 scope = the typed OAuth-credential seam only (`oauth/types.py`), the vendored
+Phase 0 scope = the typed OAuth-credential seam only (`outbound_credentials/types.py`), the vendored
 `Result` (`result.py`), and the basedpyright match-exhaustiveness spike
 (`_spike_exhaustiveness.py`). No transport, registry, or CI/semgrep/composition-root
 infrastructure — those land in Phase 2 (S0).
