+++ litellm/proxy/gateway/mcp/outbound_credentials/httpx_auth.py
